feat: get pricing in backend
