+++ src/app/stripe/route.tsx
@@ -1,13 +1,18 @@
-import { NextApiRequest } from "next";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import Stripe from "stripe";
 import { GetStripeCheckoutSessionResponse } from "./types";
+import { calculatePrice } from "../../utils/calculatePrice";
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
   apiVersion: "2022-11-15",
 });
 
-export async function GET(req: NextApiRequest, res: NextResponse) {
+export async function POST(req: NextRequest, res: NextResponse) {
+  const participants = await req.json();
+
+  const price = calculatePrice(participants);
+  console.log("price", price);
+
   const params: Stripe.Checkout.SessionCreateParams = {
     line_items: [
       {
