--- src/utils/fetcher.tsx
@@ -1,2 +1,0 @@
-// @ts-ignore
-export const fetcher = (...args) => fetch(...args).then((res) => res.json());
